weitere Fehlerbehebungen vorgenommen

+++ frontend/src/components/household/Settings.js
@@ -33,9 +33,9 @@ class Settings extends Component {
       groceries: [],
       measures: [],
       fridgeId: this.props.fridgeId,
-      showAlertGroceryDelete: false,
-      showAlertMeasureDelete: false,
       showAlertEdit: false,
+      showAlertMeasureDelete: false,
+      showAlertGroceryDelete: false,
       popupGroceryOpen: false,
       popupMeasureOpen: false,
       selectedGrocery: null,
@@ -66,12 +66,6 @@ class Settings extends Component {
     }
   }
 
-  handleDocumentClick = () => {
-    this.setState({
-      showAlertGroceryDelete: false,
-      showAlertMeasureDelete: false,
-    });
-  };
 
   handleClick = (e) => {
     const form = e.target.closest("form");
@@ -113,6 +107,68 @@ class Settings extends Component {
     this.setState({
       measures: measures,
     });
+  };
+
+  editGrocery = (groceryId) => {
+    const selectedGrocery = this.state.groceries.find(
+      (grocery) => grocery.id === groceryId
+    );
+    this.setState({
+      popupGroceryOpen: true,
+      selectedGrocery,
+    });
+  };
+
+  editMeasure = (measureId) => {
+    const selectedMeasure = this.state.measures.find(
+      (measure) => measure.id === measureId
+    );
+    this.setState({
+      popupMeasureOpen: true,
+      selectedMeasure,
+    });
+  };
+
+  updateGrocery = async (grocery) => {
+    try {
+      const updatedGrocery = await SmartFridgeAPI.getAPI().updateGrocery(
+        grocery
+      );
+      this.props.refreshGroceryList();
+      this.setState((prevState) => ({
+        groceries: prevState.groceries.map((g) =>
+          g.id === updatedGrocery.id ? updatedGrocery : g
+        ),
+        popupGroceryOpen: false,
+        selectedGrocery: null,
+      }));
+    } catch (error) {
+      this.setState({ showAlertEdit: true });
+    }
+  };
+
+  updateMeasure = async (measure) => {
+    try {
+      const updatedMeasure = await SmartFridgeAPI.getAPI().updateMeasure(
+        measure
+      );
+      this.props.refreshGroceryList();
+      this.setState((prevState) => ({
+        measures: prevState.measures.map((m) =>
+          m.id === updatedMeasure.id ? updatedMeasure : m
+        ),
+        popupMeasureOpen: false,
+        selectedMeasure: null,
+      }));
+    } catch (error) {
+      this.setState({ showAlertEdit: true });
+    }
+  };
+
+  loadGroceryStatements = async () => {
+    const Groceries = await SmartFridgeAPI.getAPI().getGroceryStatement();
+    console.log("Groceries after fetch ===>", Groceries);
+    this.setState({ GroceryList: Groceries });
   };
 
   deleteGrocery = async (groceryId) => {
@@ -159,75 +215,13 @@ class Settings extends Component {
       });
   };
 
-  editGrocery = (groceryId) => {
-    const selectedGrocery = this.state.groceries.find(
-      (grocery) => grocery.id === groceryId
-    );
-    this.setState({
-      popupGroceryOpen: true,
-      selectedGrocery,
-    });
-  };
-
-  editMeasure = (measureId) => {
-    const selectedMeasure = this.state.measures.find(
-      (measure) => measure.id === measureId
-    );
-    this.setState({
-      popupMeasureOpen: true,
-      selectedMeasure,
-    });
-  };
-
-  updateGrocery = async (grocery) => {
-    try {
-      const updatedGrocery = await SmartFridgeAPI.getAPI().updateGrocery(
-        grocery
-      );
-      this.props.refreshGroceryList()
-      this.setState((prevState) => ({
-        groceries: prevState.groceries.map((g) =>
-          g.id === updatedGrocery.id ? updatedGrocery : g
-        ),
-        popupGroceryOpen: false,
-        selectedGrocery: null,
-      }));
-    } catch (error) {
-      this.setState({ showAlertEdit: true });
-    }
-  };
-
-  // {
-  //   "id": 0,
-  //   "firstname": "string",
-  //   "lastname": "string",
-  //   "nickname": "string",
-  //   "email": "string",
-  //   "google_user_id": "string"
-  // }
-
-  updateMeasure = async (measure) => {
-    try {
-      const updatedMeasure = await SmartFridgeAPI.getAPI().updateMeasure(
-        measure
-      );
-      this.props.refreshGroceryList()
-      this.setState((prevState) => ({
-        measures: prevState.measures.map((m) =>
-          m.id === updatedMeasure.id ? updatedMeasure : m
-        ),
-        popupMeasureOpen: false,
-        selectedMeasure: null,
-      }));
-    } catch (error) {
-      this.setState({ showAlertEdit: true });
-    }
-  };
-
+ 
   handleDeleteGrocery = (groceryId) => (e) => {
-    const isExisting = this.state.GroceryList.some(grocery => grocery.grocery_id === groceryId)
+    const isExisting = this.state.GroceryList.some(
+      (grocery) => grocery.grocery_id === groceryId
+    );
     if (isExisting) {
-      alert('Grocery is already in use');
+      this.setState({ showAlertGroceryDelete: true });
       return;
     }
 
@@ -235,17 +229,18 @@ class Settings extends Component {
     this.deleteGrocery(groceryId);
   };
 
-  loadGroceryStatements = async () => {
-    const Groceries = await SmartFridgeAPI.getAPI().getGroceryStatement()
-    console.log('Groceries after fetch ===>', Groceries)
-    this.setState({ GroceryList: Groceries });
-  }
-
   handleDeleteMeasure = (measureId) => (e) => {
+    const isExisting = this.state.GroceryList.some(
+      (measure) => measure.unit_id === measureId
+    );
+    if (isExisting) {
+      this.setState({ showAlertMeasureDelete: true });
+      return;
+    }
+
     e.stopPropagation();
     this.deleteMeasure(measureId);
   };
-
   handleEditGrocery = (groceryId) => (e) => {
     e.stopPropagation();
     this.editGrocery(groceryId);
@@ -274,7 +269,7 @@ class Settings extends Component {
       popupMeasureOpen,
       selectedGrocery,
       selectedMeasure,
-      GroceryList,
+      showAlertEdit,
     } = this.state;
 
     return (
@@ -318,7 +313,6 @@ class Settings extends Component {
                   Lebensmittel bearbeiten{" "}
                 </Typography>
               </AccordionSummary>
-
               <AccordionDetails
                 sx={{ display: "flex", justifyContent: "center", m: 0, p: 0 }}
               >
@@ -326,6 +320,7 @@ class Settings extends Component {
                   sx={{
                     width: { xs: "100%", sm: "600px" },
                     marginBottom: "30px",
+                    mx: "5px",
                     display: "flex",
                     flexDirection: "column",
                     alignItems: "center",
@@ -334,6 +329,9 @@ class Settings extends Component {
                   <AlertComponent
                     showAlert={showAlertGroceryDelete}
                     alertType="SettingsGroceryDelete"
+                    onClose={() =>
+                      this.setState({ showAlertGroceryDelete: false })
+                    }
                   />
                   {groceries.map((grocery) => (
                     <ListItem
@@ -413,6 +411,7 @@ class Settings extends Component {
                   sx={{
                     width: { xs: "100%", sm: "600px" },
                     marginBottom: "30px",
+                    mx: "5px",
                     display: "flex",
                     flexDirection: "column",
                     alignItems: "center",
@@ -420,7 +419,10 @@ class Settings extends Component {
                 >
                   <AlertComponent
                     showAlert={showAlertMeasureDelete}
-                    alertType="SettingsMeasureDelete"
+                    alertType="SettingsEdit"
+                    onClose={() =>
+                      this.setState({ showAlertMeasureDelete: false })
+                    }
                   />
                   {measures.map((measure) => (
                     <ListItem
@@ -535,8 +537,9 @@ class Settings extends Component {
                     : "Einheit bearbeiten"}
                 </Typography>
                 <AlertComponent
-                  showAlert={this.state.showAlertEdit}
+                  showAlert={showAlertEdit}
                   alertType="SettingsEdit"
+                  onClose={() => this.setState({ showAlertEdit: false })}
                 />
                 <Box
                   sx={{
